% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunction.R
\name{bindingSiteDetailsPlot}
\alias{bindingSiteDetailsPlot}
\title{Plot signal coverage of selected ranges}
\usage{
bindingSiteDetailsPlot(
  object,
  plotIdx,
  flankPos,
  shiftPos = NULL,
  mergeReplicates = FALSE,
  autoscale = FALSE
)
}
\arguments{
\item{object}{a \link{BSFDataSet} object}

\item{plotIdx}{integer, index of the range to plot}

\item{flankPos}{numeric, number of nucleotides by which the plotting frame is
symmetrically extended}

\item{shiftPos}{numeric, nucleotide positions by which the current plotting
range should be shifted}

\item{mergeReplicates}{boolean, if replicates should be merge per
condition (TRUE) or if every replicates should be shown seperately (FALSE)}

\item{autoscale}{boolean, if y-axis should be scaled to the maximum for all
replicates (TRUE), or not (FALSE)}
}
\value{
an object of class \code{GVIZ}
}
\description{
Function plots the coverage of the CLIP data in the signal slot and plots it
as coverage. The plot is centered around a given binding site, which can be
selected by an index.
}
\examples{
# load data
files <- system.file("extdata", package="BindingSiteFinder")
load(list.files(files, pattern = ".rda$", full.names = TRUE))

bindingSiteDetailsPlot(bds, plotIdx = 3, flankPos = 10)


}
