% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{coverageOverRanges}
\alias{coverageOverRanges}
\title{Coverage function for BSFDataSet objects}
\usage{
coverageOverRanges(
  object,
  merge = TRUE,
  returnType = c("GRanges", "matrix", "data.frame")
)
}
\arguments{
\item{object}{a BSFDataSet object}

\item{merge}{logical, if coverage should be merged per replicate or reported
for each nucleotide in the range individually}

\item{returnType}{one of "GRanges", "matrix" or "data.frame"}
}
\value{
an object of class specified in \code{returnType}
}
\description{
The crosslink coverage is computed for all ranges in the the given \code{BSFDataSet}
object (see \code{\link{BSFDataSet}} for details). The coverage of the crosslinks
stored in the singnal slot of the \code{BSFDataSet} is computed over the ranges
in the \code{BSFDataSet} object.
}
\details{
If \code{returnType} is set to GRanges, the coverge information is stored in the
metadata slot.
}
\examples{
# load data
csFile <- system.file("extdata", "PureCLIP_crosslink_sites_example.bed",
 package="BindingSiteFinder")
cs = rtracklayer::import(con = csFile, format = "BED")
clipFiles <- system.file("extdata", package="BindingSiteFinder")

# one experimental condition
meta = data.frame(condition = c("WT", "WT", "WT", "WT"),
clPlus = list.files(clipFiles, pattern = "plus.bw$", full.names = TRUE),
clMinus = list.files(clipFiles, pattern = "minus.bw$", full.names = TRUE))
bds = BSFDataSet(ranges = cs, meta = meta)

rng = coverageOverRanges(bds)

}
