% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{exportToBED}
\alias{exportToBED}
\title{Wrapper function to export binding sites as BED files}
\usage{
exportToBED(object, con)
}
\arguments{
\item{object}{a \code{\link{BSFDataSet}} object with stored ranges}

\item{con}{A path or URL}
}
\value{
a .bed file
}
\description{
Function that serves as a wrapper function for \code{rtracklayer::export}.
}
\examples{
# load data
files <- system.file("extdata", package="BindingSiteFinder")
load(list.files(files, pattern = ".rda$", full.names = TRUE))
\dontrun{
# export
# exportToBED(bds, con = "./myfile.bed")
}

}
\seealso{
\code{\link{BSFind}}
}
