% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{forceEqualNames}
\alias{forceEqualNames}
\title{Helper function that fixes seqnames errors in ranges and signal}
\usage{
forceEqualNames(ranges, signal)
}
\arguments{
\item{ranges}{GRanges object as it is used to build a \link{BSFDataSet}}

\item{signal}{SignalList object as it is used to build a \link{BSFDataSet}}
}
\value{
GRanges list for ranges and SignalList for signal
}
\description{
Ranges and signal list must be in the same structure as it is used by
the \link{BSFDataSet}. The function removes all ranges and signal entries
where the seqlevels are not part of both signal and ranges.
}
\details{
If only ranges are problematic, keepStandardChromosomes might solve the
problem too.
}
\examples{
# load data
files <- system.file("extdata", package="BindingSiteFinder")
load(list.files(files, pattern = ".rda$", full.names = TRUE))

rng = getRanges(bds)
sgn = getSignal(bds)

r = as.data.frame(granges(rng))
d = data.frame(seqnames = "chr6", start = 1, end = 1, width = 1, strand = "+")
rngNew = makeGRangesFromDataFrame(rbind(d, r), keep.extra.columns = TRUE)

fixed = forceEqualNames(rngNew, sgn)

bdsNew = setRanges(bds, fixed$ranges)
bdsNew = setSignal(bdsNew, fixed$signal)

}
