% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{getName}
\alias{getName}
\alias{getName,BSFDataSet-method}
\title{Accessor method for the name of the BSFDataSet object}
\usage{
getName(object, ...)

\S4method{getName}{BSFDataSet}(object)
}
\arguments{
\item{object}{a \code{BSFDataSet} object}
}
\value{
returns the name of the dataset
}
\description{
The name slot holds the name of the dataset
}
\examples{

# load data
files <- system.file("extdata", package="BindingSiteFinder")
load(list.files(files, pattern = ".rda$", full.names = TRUE))

getName(bds)

}
\seealso{
\code{\link{BSFDataSet}}
}
