% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunction.R
\name{reproducibiliyCutoffPlot}
\alias{reproducibiliyCutoffPlot}
\title{Plot to that shows how many replicates support each binding site}
\usage{
reproducibiliyCutoffPlot(
  object,
  cutoff = 0.05,
  min.crosslinks = 1,
  max.range = 20,
  ...
)
}
\arguments{
\item{object}{a BSFDataSet object}

\item{cutoff}{a vector of length = 1, or of length = levels(meta$conditions) with
a single number (between 0-1) indicating the quantile cutoff}

\item{min.crosslinks}{numeric of length = 1, defines the lower boundary for the minimum
number of crosslinks a binding site has to be supported by all replicates, regardless
of the replicate specific quantile threshold}

\item{max.range}{maximum number of crosslinks per sites that should be shown}

\item{...}{further arguments passed to ggplot}
}
\value{
a plot of type \code{ggplot2} showing the per replicate reproducibility
cutoffs based on a given quantile threshold
}
\description{
Plotting function for settings specified in \code{\link{reproducibilityFilter}}.
}
\examples{

# load data
csFile <- system.file("extdata", "PureCLIP_crosslink_sites_example.bed",
 package="BindingSiteFinder")
cs = rtracklayer::import(con = csFile, format = "BED")
clipFiles <- system.file("extdata", package="BindingSiteFinder")

# two experimental conditions
meta = data.frame(condition = factor(c("WT", "WT", "KD", "KD"), levels = c("KD", "WT")),
clPlus = list.files(clipFiles, pattern = "plus.bw$", full.names = TRUE),
clMinus = list.files(clipFiles, pattern = "minus.bw$", full.names = TRUE))
bds = BSFDataSet(ranges = cs, meta = meta)

# merge binding sites
bds <- makeBindingSites(object = bds, bsSize = 9, minWidth = 2,
minCrosslinks = 2, minClSites = 1)

# use a single quantile cutoff
reproducibiliyCutoffPlot(bds, max.range = 20, cutoff = c(0.05))

# use condition specific quantile cutoffs
reproducibiliyCutoffPlot(bds, max.range = 20, cutoff = c(0.1, 0.05))


}
\seealso{
\code{\link{reproducibilityFilter}}
}
