% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{setName}
\alias{setName}
\alias{setName,BSFDataSet-method}
\title{Setter method for the names of the BSFDataSet object
The name slot holds the name information of the dataset}
\usage{
setName(object, ...)

\S4method{setName}{BSFDataSet}(object, newName)
}
\arguments{
\item{object}{a \code{BSFDataSet} object}

\item{...}{additional arguments}

\item{newName}{a character that is the name}
}
\value{
object of type \code{\link{BSFDataSet}} with updated name
}
\description{
Setter method for the names of the BSFDataSet object
The name slot holds the name information of the dataset
}
\examples{

# load data
files <- system.file("extdata", package="BindingSiteFinder")
load(list.files(files, pattern = ".rda$", full.names = TRUE))

bdsNew = setName(bds, "test01")

}
\seealso{
\code{\link{BSFDataSet}}
}
