% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{setRanges}
\alias{setRanges}
\alias{setRanges,BSFDataSet-method}
\title{Setter method for the ranges of the BSFDataSet object
The GRanges object that holds the genomic ranges information can be replaced.}
\usage{
setRanges(object, ...)

\S4method{setRanges}{BSFDataSet}(object, newRanges)
}
\arguments{
\item{object}{a \code{BSFDataSet} object}

\item{...}{additional arguments}

\item{newRanges}{an object of type \code{GRanges}}
}
\value{
object of type \code{\link{BSFDataSet}} with updated ranges
}
\description{
Setter method for the ranges of the BSFDataSet object
The GRanges object that holds the genomic ranges information can be replaced.
}
\examples{

# load data
csFile <- system.file("extdata", "PureCLIP_crosslink_sites_example.bed",
 package="BindingSiteFinder")
cs = rtracklayer::import(con = csFile, format = "BED")
clipFiles <- system.file("extdata", package="BindingSiteFinder")

# two experimental conditions
meta = data.frame(condition = factor(c("WT", "WT", "KD", "KD"), levels = c("KD", "WT")),
clPlus = list.files(clipFiles, pattern = "plus.bw$", full.names = TRUE),
clMinus = list.files(clipFiles, pattern = "minus.bw$", full.names = TRUE))
bds = BSFDataSet(ranges = cs, meta = meta)

rng = getRanges(bds)
rng = rng + 10
bdsNew = setRanges(bds, rng)

}
\seealso{
\code{\link{BSFDataSet}}
}
