% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,BSFDataSet-method}
\title{Show method to for the BSFDataSet}
\usage{
\S4method{show}{BSFDataSet}(object)
}
\arguments{
\item{object}{a \code{BSFDataSet} object}
}
\value{
shows the current object state
}
\description{
Prints the information for each of the slots in the \code{BSFDataSet} object.
Ranges of the \code{\link{getRanges}} slot are shown, as well as the number
of crosslinks per strand \code{\link{getSignal}} and the levels of the
experimental conditions (\code{\link{getMeta}}).
}
\examples{

# load data
files <- system.file("extdata", package="BindingSiteFinder")
load(list.files(files, pattern = ".rda$", full.names = TRUE))

show(bds)

}
\seealso{
\code{\link{BSFDataSet}}
}
