% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{`[`}
\alias{`[`}
\alias{[`,BSFDataSet-method}
\title{Extract parts of BSFDataSet}
\usage{
\S4method{[}{BSFDataSet,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{BSFDataSet}

\item{i}{index to subset}

\item{j}{ANY}

\item{...}{additional arguments}

\item{drop}{TRUE}
}
\value{
an object of type \code{\link{BSFDataSet}} with ranges and signal
sub-setted by index i
}
\description{
Use index i to subset a BSFDataSet. The included ranges will be reduced
according to the index and the signal will be modified accordingly. Note
that only signal that covers the ranges after sub-setting will be retained.
This means all signal 'between' binding sites will be lost.
}
\examples{

# load data
files <- system.file("extdata", package="BindingSiteFinder")
load(list.files(files, pattern = ".rda$", full.names = TRUE))

bdsNew = bds[1:10]

}
\seealso{
\code{\link{BSFDataSet}}
}
