% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{getRanges}
\alias{getRanges}
\alias{getRanges,BSFDataSet-method}
\title{Accessor method for the ranges of the BSFDataSet object}
\usage{
getRanges(object)

\S4method{getRanges}{BSFDataSet}(object)
}
\arguments{
\item{object}{a \code{BSFDataSet} object}
}
\value{
returns the genomic ranges (GRanges) of the associated ranges
}
\description{
The ranges slot holds the genomic ranges information of the sites currently
in the object. They are encoded as a GRanges object with each binding site
having a single ranges entry.
}
\examples{

# load data
csFile <- system.file("extdata", "PureCLIP_crosslink_sites_example.bed",
 package="BindingSiteFinder")
cs = rtracklayer::import(con = csFile, format = "BED")
clipFiles <- system.file("extdata", package="BindingSiteFinder")

# two experimental conditions
meta = data.frame(condition = factor(c("WT", "WT", "KD", "KD"), levels = c("KD", "WT")),
clPlus = list.files(clipFiles, pattern = "plus.bw$", full.names = TRUE),
clMinus = list.files(clipFiles, pattern = "minus.bw$", full.names = TRUE))
bds = BSFDataSet(ranges = cs, meta = meta)

getRanges(bds)

}
\seealso{
\code{\link{BSFDataSet}}
}
