% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunction.R
\name{rangeCoveragePlot}
\alias{rangeCoveragePlot}
\title{Plot crosslink events coverage over range}
\usage{
rangeCoveragePlot(object, width, name = "Coverage Plot", ...)
}
\arguments{
\item{object}{a BSFDataSet, or a list of BSFDataSet}

\item{width}{a numeric value that defines the plotting ranges}

\item{name}{plot title}

\item{...}{further arguments passed to ggplot}
}
\value{
a plot of type \code{ggplot2} displaying the crosslink coverage over
the ranges of the given \code{\link{BSFDataSet}}
}
\description{
A diagnostic plot function that allows to check the coverage of crosslink events
over different merged regions. The cverage is shown as mean over all replicates
and conditions, with a standard deviation corridor.
}
\details{
If \code{object} is a single BSFDataObject a single coverage plot will be drawn,
whereas if it is a list of BSFDataObjects, then faceting is used to make a plot for
each list element.
}
\examples{
# load data
csFile <- system.file("extdata", "PureCLIP_crosslink_sites_example.bed",
 package="BindingSiteFinder")
cs = rtracklayer::import(con = csFile, format = "BED")
clipFiles <- system.file("extdata", package="BindingSiteFinder")

# one experimental condition
meta = data.frame(condition = c("WT", "WT", "WT", "WT"),
clPlus = list.files(clipFiles, pattern = "plus.bw$", full.names = TRUE),
clMinus = list.files(clipFiles, pattern = "minus.bw$", full.names = TRUE))
bds = BSFDataSet(ranges = cs, meta = meta)

# plotting a single object
bds <- makeBindingSites(object = bds, bsSize = 9, minWidth = 2,
minCrosslinks = 2, minClSites = 1)
rangeCoveragePlot(bds, width = 20)

# plotting multiple objects
bds1 <- makeBindingSites(object = bds, bsSize = 3, minWidth = 2,
minCrosslinks = 2, minClSites = 1, sub.chr = "chr1")
bds2 <- makeBindingSites(object = bds, bsSize = 9, minWidth = 2,
minCrosslinks = 2, minClSites = 1, sub.chr = "chr1")
l = list(`1. bsSize = 3` = bds1, `2. bsSize = 9` = bds2)
rangeCoveragePlot(l, width = 20)

}
\seealso{
\code{\link{BSFDataSet}}
}
