% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{setSignal}
\alias{setSignal}
\alias{setSignal,BSFDataSet-method}
\title{Setter method for the signal data of the BSFDataSet object}
\usage{
setSignal(object, ...)

\S4method{setSignal}{BSFDataSet}(object, newSignal)
}
\arguments{
\item{object}{a BSFDataSet object}

\item{...}{additional arguments}

\item{newSignal}{list of RLE lists}
}
\value{
an object of type \code{\link{BSFDataSet}} with updated signal
}
\description{
Signal data is loaded from the path specified in \code{\link{getMeta}} columns
"clPlus" and "clMinus" and stored as a list of RLE lists.
}
\examples{

# load data
csFile <- system.file("extdata", "PureCLIP_crosslink_sites_example.bed",
 package="BindingSiteFinder")
cs = rtracklayer::import(con = csFile, format = "BED")
clipFiles <- system.file("extdata", package="BindingSiteFinder")

# two experimental conditions
meta = data.frame(condition = factor(c("WT", "WT", "KD", "KD"), levels = c("KD", "WT")),
clPlus = list.files(clipFiles, pattern = "plus.bw$", full.names = TRUE),
clMinus = list.files(clipFiles, pattern = "minus.bw$", full.names = TRUE))
bds = BSFDataSet(ranges = cs, meta = meta)

sgn = getSignal(bds)
sgn = lapply(sgn, function(selStrand){
   lapply(selStrand, function(chrList){
       chrList[names(chrList) == "chr1"]
   })
})
bdsNew = setSignal(bds, sgn)

}
\seealso{
\code{\link{BSFDataSet}}
}
